minor changes/refactoring
