+++ app/build.gradle
@@ -1,7 +1,6 @@
 apply plugin: 'com.android.application'
 
 ext {
-    gsonVersion = '2.8.2'
     supportVersion = '27.1.1'
 }
 
@@ -27,10 +26,17 @@ dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation "com.android.support:appcompat-v7:$supportVersion"
     implementation 'com.android.support.constraint:constraint-layout:1.1.0'
-    implementation 'joda-time:joda-time:2.9.9'
+    implementation 'com.jakewharton.threetenabp:threetenabp:1.1.0'
 
-    implementation "com.google.code.gson:gson:$gsonVersion"
+    implementation "com.google.code.gson:gson:2.8.2"
     implementation 'com.squareup.okhttp3:okhttp:3.10.0'
+    //retrofit
+    implementation 'com.squareup.retrofit2:retrofit:2.2.0'
+    implementation 'com.squareup.retrofit2:converter-gson:2.2.0'
+    implementation 'com.squareup.retrofit2:adapter-rxjava:2.2.0'
+    //rxJava
+    implementation 'io.reactivex:rxjava:1.2.9'
+    implementation 'io.reactivex:rxandroid:1.2.1'
 
     implementation "com.android.support:recyclerview-v7:$supportVersion"
     implementation 'com.github.PhilJay:MPAndroidChart:v3.0.3'
