+++ app/src/main/java/com/nederlonder/wqms/LiveGraphActivity.java
@@ -16,6 +16,7 @@ import com.github.mikephil.charting.data.LineDataSet;
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
 import com.nederlonder.wqms.mock.MockDataPoint;
+import com.nederlonder.wqms.models.ChartType;
 
 import java.lang.reflect.Type;
 import java.util.ArrayList;
